farmbridge_dataset/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
faker>=24.0.0
geopy>=2.4.0
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.4.0
rich>=13.7.0
pydantic>=2.6.0
jsonschema>=4.21.0
openpyxl>=3.1.0
sqlparse>=0.5.0
